@@ -12,7 +12,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 
@@ -43,16 +43,16 @@ class BankAccountControllerTest {
     @Autowired
     private MockMvc mockMvc;
 
-    @MockBean
+    @MockitoBean
     private TransferServiceImpl transferService;
 
-    @MockBean
+    @MockitoBean
     private BankAccountRepository bankAccountRepository;
 
-    @MockBean
+    @MockitoBean
     private FileService fileService;
 
-    @MockBean
+    @MockitoBean
     private ConsumerService consumerService;
 
     @Autowired
@@ -252,7 +252,7 @@ class BankAccountControllerTest {
         doThrow(new RuntimeException("Connection failed")).when(consumerService).connect(anyString());
 
         org.junit.jupiter.api.Assertions.assertThrows(
-                org.springframework.web.util.NestedServletException.class,
+                jakarta.servlet.ServletException.class,
                 () -> mockMvc.perform(get("/api/bankaccounts/connect"))
         );
 
